Refactor Athena's Shield configuration handling
