--- app/assets/athshield/variables.athshield
@@ -1,5 +1,0 @@
-{
-  "athenaShieldActivated": false,
-  "menuVisibility": "visible",
-  "debug": "false"
-}
