+++ app/assets/extraverif/variables.json
@@ -1,0 +1,5 @@
+{
+    "extraFileVerifActivated": false,
+    "menuVisibility": "visible",
+    "debug": "false"
+}
